feat(hyprland): add monitor on DP-1
Added the Thinkpad T480 monitor description on 'DP-1', using display's
default resolution and refresh rate and automatically determine the
monitors position. This is off from the top left corner, with a scale of
1.2.

The recommended rule for quickly plugging in random monitors was not
removed. This will make any monitors automatically placed on the right
of the other(s), with its preferred resolution.

The main configuration file 'hyprland.conf' linked to
'hyprland-monitors.conf' using the source keyword.