feat(hyprland): launch waybar alongside hyprland
Used 'exec-once' keyword with the value 'waybar' in order to have Waybar
launch alongside Hyprland. If using Hyprland with 'uwsm', enable it with
the following command: 'systemctl --user enable --now waybar.service'.

As for now,  uninstalled the 'uwsm' to prevent issues as per Hyprland
documentation. This commit is related to 06d4f20.